prj/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
ddgs==9.12.0
distro==1.9.0
Django==5.2.12
django-allauth==65.15.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.1
-e git+ssh://git@github.com/gyarab/2025_wt_prj_khudanych.git@641344b67f44e2253a2de8796f2101fc4e1b03ba#egg=flag_search_core
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
groq==1.1.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
lxml==6.0.2
maturin==1.12.6
oauthlib==3.3.1
openai==2.30.0
pillow==12.1.1
polib==1.2.0
primp==1.2.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
python-dotenv==1.2.2
pytz==2026.1.post1
rdflib==7.6.0
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
soupsieve==2.8.3
SPARQLWrapper==2.0.0
sqlparse==0.5.5
tenacity==9.1.4
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Wikipedia-API==0.13.0
